Refactoring: switch from using short to int (short is too much trouble)

+++ ZadatakKrizicKruzic/src/main/java/hr/bojan/zadatakkrizickruzic/core/model/Game.java
@@ -14,7 +14,7 @@ public class Game {
 
 	private int gameId;
 	private GameStatus status;
-	private short difficulty;
+	private int difficulty;
 	private Player firstPlayer;
 	private Player secondPlayer;
 	private Player winner;
